Update data.json
There's a significant number of people avoiding details on these books, so I think for now it's best to avoid mentioning titles, connections, and appearances. We do have the "spoiler" toggle of course, but that's generally assumed to be for *published* books and these are a bit of a special case. And regardless, I generally wait until publication to update books with connections, etc. (rather than getting a head start based on previews WoBs, etc.)

I've linked the titles to the Coppermind pages, so people who WANT to dig for more information can find a little bit there.